Move QueryParameters and QueryValidator
